requirements.txt
APScheduler==3.5.0
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
dj-database-url==0.4.2
Django==1.11.8
django-modelcluster==3.1
django-taggit==0.22.1
django-treebeard==4.2.0
djangorestframework==3.6.4
flake8==3.5.0
gunicorn==19.7.1
html5lib==0.999999999
idna==2.6
mccabe==0.6.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
python-dotenv==0.7.1
pytz==2017.3
requests==2.18.4
six==1.11.0
tzlocal==1.5.1
Unidecode==0.4.21
urllib3==1.22
wagtail==1.13.1
webencodings==0.5.1
whitenoise==3.3.1
Willow==1.0
